# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Volha gateway REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "help@breezy.su"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auth/check": {
            "get": {
                "description": "проверяет пароль из cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Проверить пароль",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.SWGMessage"
                        }
                    },
                    "401": {
                        "description": "Пароль не верный",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
...
        "/api/colorphotos": {
            "put": {
                "description": "Обновляет массив фотографий для конкретного продукта и цвета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product_color_photos"
                ],
                "summary": "Обновить фотографии продукта для цвета",
                "responses": {
                    "200": {
                        "description": "успех",
                        "schema": {
                            "$ref": "#/definitions/views.SWGMessage"
                        }
                    },
                    "400": {
                        "description": "неправильный ввод",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "404": {
                        "description": "не найдено",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "ошибка в микросервисе",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет массив фотографий для конкретного продукта и цвета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product_color_photos"
                ],
                "summary": "Создать фотографии продукта для цвета",
                "parameters": [
                    {
                        "description": "Новый набор фотографий",
                        "name": "photos",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/productsRPC.ProductColorPhotos"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успех",
                        "schema": {
                            "$ref": "#/definitions/views.SWGMessage"
                        }
                    },
                    "400": {
                        "description": "неправильный ввод",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "409": {
                        "description": "уже существует",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "ошибка в микросервисе",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет записи фотографий продукта для указанного цвета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product_color_photos"
                ],
                "summary": "Удалить фотографии продукта для цвета",
                "parameters": [
                    {
                        "description": "ID продукта и цвета",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/productsRPC.ProductColorPhotosId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успех",
                        "schema": {
                            "$ref": "#/definitions/views.SWGMessage"
                        }
                    },
                    "400": {
                        "description": "неправильный ввод",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "404": {
                        "description": "не найдено",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "ошибка в микросервисе",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
...
        "/api/colorphotos/photos/get": {
            "post": {
                "description": "Возвращает массив фотографий для указанного продукта и цвета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product_color_photos"
                ],
                "summary": "Получить фотографии продукта по ID продукта и цвета",
                "parameters": [
                    {
                        "description": "ID продукта и цвета",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/productsRPC.ProductColorPhotosId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фотографий",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
...
        "/api/dictionaries/all": {
            "get": {
                "description": "Возвращает список всех доступных справочных данных: бренды, категории, материалы, страны и цвета",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionaries"
                ],
                "summary": "Получить все справочники",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/productsRPC.Dictionaries"
                        }
                    },
                    "400": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
        "/api/dictionaries/all/by-category": {
            "get": {
                "description": "Возвращает список всех доступных справочных данных: бренды, материалы, страны и цвета и размеры по определенной категории",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionaries"
                ],
                "summary": "Получить все справочники по категории",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID категории",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/productsRPC.Dictionaries"
                        }
                    },
                    "400": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
...
        "/api/product/all": {
            "get": {
                "description": "Возвращает список всех продуктов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Получить все продукты",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "start \u003e 0",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "end",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productsRPC.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
        "/api/product/filter": {
            "post": {
                "description": "Возвращает список продуктов, соответствующих заданным критериям фильтрации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Получить продукты по фильтру",
                "parameters": [
                    {
                        "description": "Параметры фильтрации",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/productsRPC.ProductFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productsRPC.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
        "/api/product/search": {
            "get": {
                "description": "Ищет продукт по ID (UUID), article (8 цифр) или названию (частичное совпадение)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Поиск продуктов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поисковый запрос",
                        "name": "promt",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/productsRPC.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "productsRPC.Brand": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "productsRPC.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "productsRPC.Color": {
            "type": "object",
            "properties": {
                "hex": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "productsRPC.Country": {
            "type": "object",
            "properties": {
                "friendly": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "productsRPC.Dictionaries": {
            "type": "object",
            "properties": {
                "brands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Brand"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Category"
                    }
                },
                "colors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Color"
                    }
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Country"
                    }
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Material"
                    }
                },
                "max_depth": {
                    "type": "integer"
                },
                "max_height": {
                    "type": "integer"
                },
                "max_price": {
                    "type": "integer"
                },
                "max_width": {
                    "type": "integer"
                },
                "min_depth": {
                    "type": "integer"
                },
                "min_height": {
                    "type": "integer"
                },
                "min_price": {
                    "type": "integer"
                },
                "min_width": {
                    "type": "integer"
                }
            }
        },
        "productsRPC.Material": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "productsRPC.Product": {
            "type": "object",
            "properties": {
                "article": {
                    "type": "string"
                },
                "brand": {
                    "$ref": "#/definitions/productsRPC.Brand"
                },
                "category": {
                    "$ref": "#/definitions/productsRPC.Category"
                },
                "colors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Color"
                    }
                },
                "country": {
                    "$ref": "#/definitions/productsRPC.Country"
                },
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Material"
                    }
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "seems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productsRPC.Product"
                    }
                },
                "title": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "productsRPC.ProductColorPhotos": {
            "type": "object",
            "properties": {
                "color_id": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
        "productsRPC.ProductColorPhotosId": {
            "type": "object",
            "properties": {
                "color_id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
        "productsRPC.ProductFilter": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "colors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_depth": {
                    "type": "integer"
                },
                "max_height": {
                    "type": "integer"
                },
                "max_price": {
                    "type": "integer"
                },
                "max_width": {
                    "type": "integer"
                },
                "min_depth": {
                    "type": "integer"
                },
                "min_height": {
                    "type": "integer"
                },
                "min_price": {
                    "type": "integer"
                },
                "min_width": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "sort_by": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "string"
                }
            }
        },
...
        "views.SWGError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "views.SWGMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "some info"
                }
            }
        }
    }
}